+++ lib/DOM/element.d.ts
@@ -1,1 +1,1 @@
-export declare function makeEl(tag: string, classList?: Array<string>): HTMLElement;
+export declare function makeEl(tag: string, classList?: Array<string>, type?: string): HTMLElement;