+++ lib/interface.d.ts
@@ -3,6 +3,8 @@ export interface ListItem {
     text: string;
     icon: Array<string>;
     closed: boolean;
+    checkBox: boolean;
+    open: boolean;
     items: Array<ListItem> | null;
 }
 export interface Options {
@@ -16,5 +18,6 @@ export interface Options {
     subListClass: Array<string>;
     listSubClass: Array<string>;
     listSubArrowClass: Array<string>;
+    checkBoxClass: Array<string>;
     itemClass: Array<string>;
 }