+++ src/DOM/element.ts
@@ -1,7 +1,7 @@
 import { addClass } from "../helpers/addClass";
 
-export function makeEl(tag:string, classList: Array<string> = []) {
-    let el = document.createElement(tag);
-
+export function makeEl(tag:string, classList: Array<string> = [], type:string = "") {
+    let el = document.createElement(tag) as HTMLInputElement;
+    if(type.length > 0) el.type = type;
     return addClass(el, classList);
 }