+++ src/callbacks.ts
@@ -2,7 +2,8 @@ import { EventType } from "./enum";
 //-----------------------------------------------------------------------------------------------
 //CALLBACKS LIST AND FIRE FUNCTION
 //-----------------------------------------------------------------------------------------------
-let _callbacks:Array<CallableFunction> = [];        //Calback array
+class CB {
+    private _callbacks:Array<CallableFunction> = [];        //Calback array
 /**
  * Fire callback according EventType
  * @param {EventType} event - Type of event
@@ -12,9 +13,9 @@ let _callbacks:Array<CallableFunction> = [];        //Calback array
  * @param {any} data3 - Forth data for callback
  * @param {any} data4 - Fifth data for callback
  */
-export function fire(event:EventType, data:any = false, data1:any = false, data2:any = false, data3:any = false, data4:any = false):void {
+public fire(event:EventType, data:any = false, data1:any = false, data2:any = false, data3:any = false, data4:any = false):void {
     //@ts-ignore
-    let callback = _callbacks[event];
+    let callback = this._callbacks[event];
     if(callback) {
         callback(data, data1, data2, data3, data4);
     }
@@ -27,14 +28,17 @@ export function fire(event:EventType, data:any = false, data1:any = false, data2
  * @param {EventType} event - Event type for callback
  * @param callback - Callback function what will be fired
  */
-export function on(event:EventType, callback:CallableFunction) {
-    const values = Object.values(EventType);
-    if (values.includes(event as unknown as EventType)) {
-        //Push function to callbacks list
-        //@ts-ignore
-        _callbacks[event] = callback;
-    }
-    else {
-        console.log("Wrong event type. Pls check it again", event);
-    }
-} 
+public on(event:EventType, callback:CallableFunction) {
+        const values = Object.values(EventType);
+        if (values.includes(event as unknown as EventType)) {
+            //Push function to callbacks list
+            //@ts-ignore
+            this._callbacks[event] = callback;
+        }
+        else {
+            console.log("Wrong event type. Pls check it again", event);
+        }
+    } 
+}
+
+export { CB };