+++ src/enum.ts
@@ -1,5 +1,15 @@
 export enum EventType {
     edit = "edit",
-    click = "click",
-    dblclick = "dblclick"
+    itemclick = "item.click",
+    itemdblclick = "item.dblclick",
+    folderclick = "folder.click",
+    folderdblclick = "folder.dblclick",
+    itemcheckbox = "item.checkbox",
+    foldercheckbox = "folder.checkbox"
+}
+
+export enum ItemType {
+    item = "item",
+    folder = "folder",
+    both = "both"
 }